+++ backend/src/modules/activity/controller.ts
@@ -1,15 +1,15 @@
 import express, { Request, Response, Router } from "express";
 import { Pool } from "pg";
-import { PostRepository } from "./repository/activity";
+import { PostRepository } from "./repository/post";
 import { CreatePostDTO, UpdatePostDTO } from "./dtos";
 
 export class PostController {
   private router: Router;
-  private activityRepository: PostRepository;
+  private postRepository: PostRepository;
 
   constructor(pool: Pool) {
     this.router = express.Router();
-    this.activityRepository = new PostRepository(pool);
+    this.postRepository = new PostRepository(pool);
     this.initializeRoutes();
   }
 
@@ -23,8 +23,8 @@ export class PostController {
 
   getAll = async (req: Request, res: Response) => {
     try {
-      const activities = await this.activityRepository.getPosts();
-      res.json(activities);
+      const posts = await this.postRepository.getPosts();
+      res.json(posts);
     } catch (error) {
       res.status(500).json({ message: "Erro Interno!" });
     }
@@ -34,7 +34,7 @@ export class PostController {
     try {
       const { title, description, date } = req.body;
       const createPostDTO: CreatePostDTO = { title, description, date };
-      await this.activityRepository.createPost(createPostDTO);
+      await this.postRepository.createPost(createPostDTO);
       res.status(201).json({ message: "Atividade Criada com sucesso" });
     } catch (error) {
       res.status(500).json({ message: "Erro Interno!" });
@@ -44,11 +44,11 @@ export class PostController {
   getById = async (req: Request, res: Response) => {
     try {
       const { id } = req.params;
-      const activity = await this.activityRepository.getPostById(id);
-      if (!activity) {
+      const post = await this.postRepository.getPostById(id);
+      if (!post) {
         res.status(404).json({ message: "Atividade nao encontrada" });
       } else {
-        res.json(activity);
+        res.json(post);
       }
     } catch (error) {
       res.status(500).json({ message: "Erro Interno!" });
@@ -65,7 +65,7 @@ export class PostController {
         description,
         date,
       };
-      await this.activityRepository.updatePost(updatePostDTO);
+      await this.postRepository.updatePost(updatePostDTO);
       res.json({ message: "Atividade Atualizada com sucesso" });
     } catch (error) {
       res.status(500).json({ message: "Erro Interno!" });
@@ -75,7 +75,7 @@ export class PostController {
   delete = async (req: Request, res: Response) => {
     try {
       const { id } = req.params;
-      await this.activityRepository.deletePost(id);
+      await this.postRepository.deletePost(id);
       res.json({ message: "Atividade Deletada com sucesso" });
     } catch (error) {
       res.status(500).json({ message: "Erro Interno!" });
