+++ backend/src/modules/activity/dtos/get-post.dto.ts
@@ -1,5 +1,5 @@
 import { PostEntity } from "../../../@types/entities";
 
 export interface GetPostsDTO {
-  activities: PostEntity[];
+  posts: PostEntity[];
 }
