+++ backend/src/modules/activity/dtos/index.ts
@@ -1,2 +1,2 @@
-export * from './create-activity.dto';
-export * from './update-activity.dto';
+export * from "./create-post.dto";
+export * from "./update-post.dto";
