+++ backend/src/modules/posts/dtos/get-post.dto.ts
@@ -1,5 +1,5 @@
 import { PostEntity } from "../../../@types/entities";
 
 export interface GetPostDTO {
-  activities: PostEntity[];
+  posts: PostEntity[];
 }
