+++ backend/src/modules/user/dtos/get-user.dto.ts
@@ -1,5 +1,5 @@
 import { UserEntity } from "../../../@types/entities";
 
 export interface GetUserDTO {
-  activities: UserEntity[];
+  posts: UserEntity[];
 }
